Revert "Moved modal component to UI folder"
