+++ components/fighter/fighter-image-edit-modal.tsx
@@ -6,7 +6,7 @@ import { Button } from '@/components/ui/button';
 import { createClient } from '@/utils/supabase/client';
 import { useToast } from '@/components/ui/use-toast';
 import { updateFighterImage } from '@/app/actions/update-fighter-image';
-import Modal from '@/components/ui/modal';
+import Modal from '@/components/modal';
 
 interface FighterImageEditModalProps {
   isOpen: boolean;
