+++ components/gang/add-fighter.tsx
@@ -3,7 +3,7 @@
 import React, { useState, useEffect } from 'react';
 import { Input } from '../ui/input';
 import { Button } from '../ui/button';
-import Modal from '@/components/ui/modal';
+import Modal from '@/components/modal';
 import { FighterType } from '@/types/fighter-type';
 import { useToast } from "@/components/ui/use-toast";
 import { fighterClassRank } from "@/utils/fighterClassRank";
