+++ components/gang/gang-logs.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React, { useState, useEffect } from 'react';
-import Modal from '@/components/ui/modal';
+import Modal from '@/components/modal';
 import { List, ListColumn, ListAction } from '@/components/ui/list';
 
 interface GangLog {
